+++ CHANGELOG.md
@@ -34,6 +34,10 @@
   `use_nested_configs` configuration key are unaffected.  
   [JP Simard](https://github.com/jpsim)
 
+* The `use_nested_configs` configuration key has been deprecated and its value
+  is now ignored. Nested configuration files are now always considered.  
+  [JP Simard](https://github.com/jpsim)
+
 ##### Enhancements
 
 * `swiftlint lint` now accepts an optional `--reporter` parameter which